fix: unit test fail,  grouping review moved under migration, inquirerpy dependency added
